refactor: measurement handling and extract measurement conversions utility
